Fixing the order of arguments in _fix_code_with_llm to be consistent with how the tool is being called (#898)
All of the calls to _fix_code_with_llm were passing the arguments in the
order `content, file_path, error description`. The function actually had
an order of `code, error description, context`. This resulted in the
`fix_code` prompt not actually having the error message when prompting
the LLM and resulted in many fix_code calls that didn't have an effect.