Faster Web Dev Agent Iterations (#1064)
The current web dev agent takes a long time to debug due to the fact
that context continues to grow as we iterate on a design. I anecdotally
have an example of an application taking 24 minutes to complete 5
different plans to eventually have a finished application with no
errors.

This PR adds the ability for the agent to summarize conversation
history. This is done through exposing a
`summarize_conversation_history(max_history_tokens)` API in the ChatSDK.
When this is called, if the current number of tokens in the saved
history is greater than `max_history_tokens`, the agent will summarize
all of the saved history and then replace the saved history with that
summarization. In the web dev agent, we call this function as we start
new debugging iterations.

There are also some misc changes such as:
1. Changing an issue in the template typing that was creating a
validation error for one of the applications the agent couldn't fix.
2. Adding prompting and functionality for the exact error string to be
sent to the `fix_code` tool, increasing the debugging capabilities of
the agent.

This has been tested to work well across the prompts that we are using
to evaluate.