Limiting the code agent router to only use typescript (#1128)
The code agent only supports typescript at this time. We now report an
error if the router detects that another language/framework is
requested. If the router is not sure what language the user wants it
will also assume it is typescript as that is currently the only language
supported by the code agent.